feat: use Canvas API and SVG graphics
Using Canvas API, all the computations to overlap the frame are done
client-side, providing a much faster and efficient experience, both for
users and for the convenience of the server.

Moreover, using SVGs instead of bitmap images allows to dynamically
change at run-time the color of the frame, and this is useful to
drastically drop the used bandwith and allows to customize the frames at
will.